view tests refactoring.
